# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChemAI Model Extractor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects visible to the current user
      description: >-
        If authenticated, returns the user's own projects plus all public projects.
        If not authenticated, returns only public projects.
      responses:
        "200":
          description: List of projects with summary stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with all its source documents and extractions
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with nested data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (cascades to all sources and extractions)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{projectId}/visibility:
    patch:
      operationId: updateProjectVisibility
      tags: [projects]
      summary: Update project visibility (owner only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectVisibilityInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /projects/{projectId}/sources:
    post:
      operationId: addSourceDocument
      tags: [sources]
      summary: Add a source document (raw text or pasted PDF text) to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSourceDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDocument"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{projectId}/extractions:
    post:
      operationId: createExtraction
      tags: [extractions]
      summary: Run an extraction on the most recent source document and save the result
      description: |
        Uses the active LLM provider (mock by default). Persists the model card, equations,
        variables, parameters, assumptions, limitations, and ODE template.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExtractionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCard"
        "400":
          description: No source document available to extract from
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{projectId}/model-card:
    get:
      operationId: getModelCardByProject
      tags: [extractions]
      summary: Get the latest model card for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest model card with all related artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCard"
        "404":
          description: Project or model card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{projectId}/export:
    get:
      operationId: exportProject
      tags: [export]
      summary: Export the project and all stored data as a single JSON document
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exported project bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExport"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UpdateProjectVisibilityInput:
      type: object
      properties:
        visibility:
          type: string
          enum: [private, public]
      required: [visibility]
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        ownerId:
          type: ["string", "null"]
          description: User ID of the project owner. Null for legacy/demo projects.
        visibility:
          type: string
          enum: [private, public]
          description: Who can view this project. Private projects are only accessible to their owner.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, description, ownerId, visibility, createdAt, updatedAt]
    ProjectSummary:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            sourceDocumentCount:
              type: integer
            extractionCount:
              type: integer
            latestExtractionTitle:
              type: ["string", "null"]
          required:
            - sourceDocumentCount
            - extractionCount
            - latestExtractionTitle
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            sourceDocuments:
              type: array
              items:
                $ref: "#/components/schemas/SourceDocument"
            extractions:
              type: array
              items:
                $ref: "#/components/schemas/ExtractionSummary"
          required: [sourceDocuments, extractions]
    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required: [name]
    SourceDocument:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        kind:
          type: string
          enum: [text, pdf]
        filename:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, kind, filename, content, createdAt]
    AddSourceDocumentInput:
      type: object
      properties:
        kind:
          type: string
          enum: [text, pdf]
        filename:
          type: ["string", "null"]
        content:
          type: string
          minLength: 1
      required: [kind, content]
    CreateExtractionInput:
      type: object
      properties:
        sourceDocumentId:
          type: ["integer", "null"]
          description: Optional — if omitted, uses the most recent source document for the project.
        provider:
          type: string
          enum: [mock, openai, gemini, auto]
          description: >-
            AI provider to use for extraction. "auto" (default) uses the
            fallback chain: OpenAI → Gemini → Mock depending on which keys
            are configured. "mock" always uses deterministic mock output.
...
    Assumption:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        kind:
          type: string
          enum: [assumption, limitation]
        text:
          type: string
        sourceQuote:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, kind, text, sourceQuote, confidence, editedByUser]
...
    ModelCard:
      type: object
      properties:
        extraction:
          type: object
          properties:
            id:
              type: integer
            projectId:
              type: integer
            sourceDocumentId:
              type: ["integer", "null"]
            providerUsed:
              type: string
              enum: [mock, openai, gemini]
            status:
              type: string
              enum: [pending, ready, failed]
            modelCardTitle:
              type: string
            domain:
              type: string
            systemDescription:
              type: string
            problemStatement:
              type: string
            odeTemplate:
              type: string
            rawExtractionJson:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Full validated provider output matching the canonical
                ExtractionResultSchema (paper_title_or_topic, system_type,
                process_description, state_variables, parameters, equations,
                assumptions, limitations, model_card). Null for legacy rows
                created before this column existed; clients should fall back
                to the normalized tables in that case.
            providerModel:
              type: string
              description: >-
                Exact model identifier used by the provider
                (e.g. "gpt-4o", "gemini-1.5-flash", "mock").
                Empty string for legacy rows.
            systemPrompt:
              type: string
              description: >-
                System prompt sent verbatim to the AI provider.
                Contains only instructional text — NO API keys or secrets.
                Empty string for mock provider, legacy rows, and public share views.
            promptTemplateSummary:
              type: string
              description: One-line description of the extraction prompt template purpose.
            rawProviderResponse:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Raw provider response BEFORE JSON repair and Zod validation.
                Distinct from rawExtractionJson (post-validation canonical result).
                Null for mock provider (no API call made), legacy rows, and public share views.
            repairStatus:
              type: string
              enum: [not_needed, repaired, failed]
              description: Whether the provider response required JSON repair before validation.
            validationErrors:
              type: ["string", "null"]
              description: >-
                Zod validation error details if the extraction failed or
                required repair. Null when the extraction succeeded cleanly.
            tokenUsage:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Token count and estimated cost metadata from the provider.
                Shape varies by provider (OpenAI vs Gemini). Null when
                unavailable (mock provider always returns null).
            modelType:
              type: string
              enum:
                - chemostat
                - batch_reactor
                - fed_batch
                - cstr
                - gas_liquid_transfer
                - microalgae_pbr
                - generic_ode
              description: >-
                Auto-detected model type from the rule-based domain classifier.
                "generic_ode" is the safe default for unknown models and
                legacy rows.
            modelTypeConfidence:
              type: number
              format: float
              description: >-
                Classifier confidence score in [0, 1].
                Computed as score / (score + 10) from keyword evidence.
                0 means no domain keywords were found.
            modelTypeMatchedKeywords:
              type: array
              items:
                type: string
              description: >-
                Keywords and phrases from the source text / extracted fields
                that contributed to the auto-classification score.
            modelTypeOverride:
              type: ["string", "null"]
              enum:
                - chemostat
                - batch_reactor
                - fed_batch
                - cstr
                - gas_liquid_transfer
                - microalgae_pbr
                - generic_ode
                - null
              description: >-
                User-supplied model type override. When non-null, takes
                precedence over modelType in all frontend displays. Null means
                "use the classifier result".
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - projectId
            - sourceDocumentId
            - providerUsed
            - status
            - modelCardTitle
            - domain
            - systemDescription
            - problemStatement
            - odeTemplate
            - rawExtractionJson
            - providerModel
            - systemPrompt
            - promptTemplateSummary
            - rawProviderResponse
            - repairStatus
            - validationErrors
            - tokenUsage
            - modelType
            - modelTypeConfidence
            - modelTypeMatchedKeywords
            - modelTypeOverride
            - createdAt
            - updatedAt
        equations:
          type: array
          items:
            $ref: "#/components/schemas/Equation"
        variables:
          type: array
          items:
            $ref: "#/components/schemas/Variable"
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/Parameter"
        assumptions:
          type: array
          items:
            $ref: "#/components/schemas/Assumption"
      required: [extraction, equations, variables, parameters, assumptions]
    ProjectExport:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        sourceDocuments:
          type: array
          items:
            $ref: "#/components/schemas/SourceDocument"
        modelCards:
          type: array
          items:
            $ref: "#/components/schemas/ModelCard"
        exportedAt:
          type: string
          format: date-time
      required: [project, sourceDocuments, modelCards, exportedAt]